+++ example/lib/features/webview_handoff/presentation/webview_handoff_page.dart
@@ -136,4 +136,3 @@ class _MissingConfigCard extends StatelessWidget {
     );
   }
 }
-